test: adds tests and skeletons for e2e testing
I realized I should've broken this commit up a bit more. Oops.
So the `app.e2e-spec.ts` file should now contain tests for `AppController`
`ThrottlerController` and the setup for `DefaultController` and
`LimitController`. I figured each of these controllers would test a
specific segment of the application and what it does. The `AppController`
gives good tests to the `@Throttle()` decorator and to the `ignoreRoutes`
option. The `ThrottlerController` tests the `@Throttle()` decorator on
both the class level and the route level ensuring we can override the
route level. The `DefaultController` will be for not using the decorator
but still getting the headers to show up as expected. The
`LimitController` will be for ensuring that once we pass limits, we get
the expected return values in the header.

I've made a few other updates to the `ThrottlerGuard` (allowing for
class metadata and taking into account the record that is about to be
added), the `ThrottlerCoreModule` (changing the provided token), and the
`ThrottlerOptions` interface (using an instance instead of a class) to
make things easier for us as we go forward. There's also now a
`test:e2e:dev` command to run `jest` in watch mode.