feat(guard): clean up guard and update docs
The guard no longer handles ws and gql by default. This is done
to allow for more extensibility by others and to allow for others
to create an exception class specific to their needs. The optional
requre that was being used for the `WsException` does not work
to keep the class instance, so this is the cleanest thing I can
currently come up with